logic the validator and RequiredArgsContructor

+++ src/main/java/com/miguelsouza/libraryapi/service/AutorService.java
@@ -1,8 +1,11 @@
 package com.miguelsouza.libraryapi.service;
 
+import com.miguelsouza.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import com.miguelsouza.libraryapi.model.Autor;
 import com.miguelsouza.libraryapi.repository.AutorRepository;
+import com.miguelsouza.libraryapi.repository.LivroRepository;
 import com.miguelsouza.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -11,17 +14,15 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository repository;
     private final AutorValidator validator;
-
-    public AutorService(AutorRepository repository, AutorValidator validator) {
-        this.repository = repository;
-        this.validator = validator;
-    }
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor) {
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -29,6 +30,7 @@ public class AutorService {
         if (autor.getId() == null) {
             throw new IllegalArgumentException("Para atualizar, é necessario um autor já salvo");
         }
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -37,6 +39,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Autor possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -55,5 +60,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
